Feat (ListeningHistory): Refactor user listening history retrieval and update logic; enhance DTO with additional fields

+++ backend/src/main/java/com/example/demo/model/ListeningHistory.java
@@ -27,5 +27,5 @@ public class ListeningHistory {
     private Song song;
 
     private Timestamp playedAt;
-    private int listenDuration;
+    private long listenDuration;
 }
